src/components/tools: InfographicsTool switch on media
